+++ tools/integration_tests/readonly/read_test.go
@@ -19,7 +19,6 @@ import (
 	"fmt"
 	"os"
 	"path"
-	"strings"
 	"syscall"
 	"testing"
 
@@ -58,13 +57,10 @@ func TestReadFileFromSubDirectory(t *testing.T) {
 	}
 }
 
-func checkIfNonExistentFileOpenFails(filePath string, t *testing.T) {
+func checkIfNonExistentFileFailsToOpen(filePath string, t *testing.T) {
 	file, err := os.OpenFile(filePath, os.O_RDONLY|syscall.O_DIRECT, setup.FilePermission_0600)
 
-	// It will throw an error no such file or directory.
-	if !strings.Contains(err.Error(), "no such file or directory") {
-		t.Errorf("Throwing incorrect error.")
-	}
+	checkErrorForObjectNotExist(err, t)
 
 	if err == nil {
 		t.Errorf("Nonexistent file opened to read.")
@@ -75,11 +71,11 @@ func checkIfNonExistentFileOpenFails(filePath string, t *testing.T) {
 func TestReadNonExistentFile(t *testing.T) {
 	filePath := path.Join(setup.MntDir(), FileNotExist)
 
-	checkIfNonExistentFileOpenFails(filePath, t)
+	checkIfNonExistentFileFailsToOpen(filePath, t)
 }
 
 func TestReadNonExistentFileFromSubDirectory(t *testing.T) {
 	filePath := path.Join(setup.MntDir(), DirectoryNameInTestBucket, FileNotExist)
 
-	checkIfNonExistentFileOpenFails(filePath, t)
+	checkIfNonExistentFileFailsToOpen(filePath, t)
 }
