+++ tools/integration_tests/readonly/rename_object_test.go
@@ -17,83 +17,12 @@
 package readonly_test
 
 import (
-	"io"
 	"os"
-	"os/exec"
 	"path"
-	"strings"
-	"syscall"
 	"testing"
 
 	"github.com/googlecloudplatform/gcsfuse/tools/integration_tests/setup"
 )
-
-// Copy srcFile in testBucket/Test/b/b.txt destination.
-func checkIfFileCopyFailed(srcFilePath string, t *testing.T) {
-	source, err := os.OpenFile(srcFilePath, syscall.O_DIRECT, setup.FilePermission_0600)
-	if err != nil {
-		t.Errorf("Error in the opening file: %v", err)
-	}
-
-	// Checking if destination object exist.
-	copyFile := path.Join(setup.MntDir(), DirectoryNameInTestBucket, SubDirectoryNameInTestBucket, "b.txt")
-	if _, err := os.Stat(copyFile); err != nil {
-		t.Errorf("Copied file %s is not present", copyFile)
-	}
-
-	destination, err := os.OpenFile(copyFile, syscall.O_DIRECT, setup.FilePermission_0600)
-	if err != nil {
-		t.Errorf("File %s opening error: %v", destination.Name(), err)
-	}
-	defer destination.Close()
-
-	// File copying with io.Copy() utility.
-	_, err = io.Copy(destination, source)
-	// Throwing an error "copy_file_range: bad file descriptor"
-	if err == nil {
-		t.Errorf("File copied in read-only file system.")
-	}
-}
-
-// Copy testBucket/Test1.txt to testBucket/Test/b/b.txt
-func TestCopyFile(t *testing.T) {
-	file := path.Join(setup.MntDir(), FileNameInTestBucket)
-
-	checkIfFileCopyFailed(file, t)
-}
-
-// Copy testBucket/Test/a.txt to testBucket/Test/b/b.txt
-func TestCopyFileFromSubDirectory(t *testing.T) {
-	file := path.Join(setup.MntDir(), DirectoryNameInTestBucket, FileNameInDirectoryTestBucket)
-
-	checkIfFileCopyFailed(file, t)
-}
-
-func checkIfDirCopyFailed(srcDirPath string, newDir string, t *testing.T) {
-	cmd := exec.Command("cp", "--recursive", srcDirPath, newDir)
-	err := cmd.Run()
-
-	// Throwing an  exit status 1
-	if err == nil {
-		t.Errorf("Dir copied in read-only file system.")
-	}
-}
-
-// Copy testBucket/Test to testBucket/Test/b
-func TestCopyDir(t *testing.T) {
-	srcDir := path.Join(setup.MntDir(), DirectoryNameInTestBucket)
-	destDir := path.Join(setup.MntDir(), DirectoryNameInTestBucket, SubDirectoryNameInTestBucket)
-
-	checkIfDirCopyFailed(srcDir, destDir, t)
-}
-
-// Copy testBucket/Test/b to testBucket/Test
-func TestCopySubDirectory(t *testing.T) {
-	srcDir := path.Join(setup.MntDir(), DirectoryNameInTestBucket, SubDirectoryNameInTestBucket)
-	destDir := path.Join(setup.MntDir(), DirectoryNameInTestBucket)
-
-	checkIfDirCopyFailed(srcDir, destDir, t)
-}
 
 // Rename srcFile to Rename.txt
 func checkIfFileRenameFailed(oldObjPath string, newObjPath string, t *testing.T) {
@@ -112,10 +41,7 @@ func checkIfFileRenameFailed(oldObjPath string, newObjPath string, t *testing.T)
 		t.Errorf("File renamed in read-only file system.")
 	}
 
-	// It will throw an error read-only file system or permission denied.
-	if !strings.Contains(err.Error(), "read-only file system") && !strings.Contains(err.Error(), "permission denied") {
-		t.Errorf("Throwing incorrect error.")
-	}
+	checkErrorForReadOnlyFileSystem(err, t)
 
 	if _, err := os.Stat(oldObjPath); err != nil {
 		t.Errorf("SrcFile is deleted in read-only file system.")
